Clustering fixes and cleanup.
