+++ survos2/api/analyzer.py
@@ -1,6 +1,6 @@
 import ntpath
 from typing import List
-
+import os
 import matplotlib.patheffects as PathEffects
 import numpy as np
 import pandas as pd
@@ -1181,7 +1181,8 @@ def object_analyzer(
     with DatasetManager(src, out=None, dtype="float32", fillvalue=0) as DM:
         ds_objects = DM.sources[0]
     entities_fullname = ds_objects.get_metadata("fullname")
-    tabledata, entities_df = setup_entity_table(entities_fullname, flipxy=flipxy)
+    objects_path = ds_objects._path
+    tabledata, entities_df = setup_entity_table(os.path.join(objects_path, entities_fullname), flipxy=flipxy)
 
     entities = np.array(make_entity_df(np.array(entities_df), flipxy=False))
