+++ survos2/entity/anno/crowd.py
@@ -692,220 +692,3 @@ def generate_stacked_click_plot_data(img_data_layers, click_coords, patch_size=(
         img_shortlists = [img_shortlistA, img_shortlistB]
 
     return img_shortlists, img_titles
-
-
-# """
-# change the underlying bb rep from center + w,h,d to lbl to utr
-# resample z? (x 5?) then resize the image?
-# """
-
-
-# def generate_wide_plot_data(
-#     img_data,
-#     click_coords,
-#     wide_patch_pos=(63, 650, 650),
-#     z_depth=3,
-#     patch_size=(40, 200, 200),
-# ):
-#     img_titles = []
-#     patch_size = np.array(patch_size)
-#     sliceno, x, y = wide_patch_pos
-
-#     z_depth, p_x, p_y = patch_size
-#     w = int(p_x / 2.0)
-#     h = int(p_y / 2.0)
-#     print("x y w h, sliceno: {}".format((x, y, w, h, sliceno)))
-#     # sel_cropped_clicks.append((sliceno, x,y,w,h))
-
-#     z, x_bl, x_ur, y_bl, y_ur = int(sliceno), x - w, x + w, y - h, y + h
-#     print(z, x_bl, x_ur, y_bl, y_ur)
-
-#     slice_start = np.max([0, wide_patch_pos[0] - np.int(patch_size[0] / 2.0)])
-#     slice_end = np.min([wide_patch_pos[0] + np.int(patch_size[0] / 2.0), 165])
-
-#     print("Slice start, slice end {} {}".format(slice_start, slice_end))
-#     out_of_bounds_w = np.hstack(
-#         (
-#             np.where(orig_click_data[:, 1] >= x_ur)[0],
-#             np.where(orig_click_data[:, 1] <= x_bl)[0],
-#             np.where(orig_click_data[:, 2] >= y_ur)[0],
-#             np.where(orig_click_data[:, 2] < y_bl)[0],
-#             np.where(orig_click_data[:, 0] <= slice_start)[0],
-#             np.where(orig_click_data[:, 0] >= slice_end)[0],
-#         )
-#     )
-
-#     click_data_w = np.delete(orig_click_data, out_of_bounds_w, axis=0)
-
-#     click_data_wide_arr = np.array(click_data_w)
-#     print("Click_data_wide_arr shape: {}".format(click_data_wide_arr.shape))
-
-#     # click_data_wide_arr[:,0] = click_data_wide_arr[:,0] - x_bl
-#     # click_data_wide_arr[:,1] = click_data_wide_arr[:,0] - y_bl
-#     print("Length of original click_data {}".format(orig_click_data.shape[0]))
-#     print(
-#         "Length after deleting out of bounds clicks: {}".format(
-#             click_data_wide_arr.shape[0]
-#         )
-#     )
-
-#     sel_wide_clicks, x_coords_range, y_coords_range = generate_clicklist(
-#         click_data_wide_arr, crop_roi, slice_start, slice_end
-#     )
-
-#     if z_depth > 1:
-#         img = get_vol_in_bbox(
-#             img_data, slice_start, slice_end, int(np.ceil(y)), int(np.ceil(x)), h, w
-#         )
-#     else:
-#         img = get_img_in_bbox(img_data, sliceno, int(np.ceil(x)), int(np.ceil(y)), w, h)
-
-#     y_str = "{:10.4f}".format(y)
-#     x_str = "{:10.4f}".format(x)
-
-#     img_titles.append(x_str + " " + y_str + "\n " + "Slice no: " + str(sliceno))
-
-#     return img, img_titles, click_data_wide_arr
-
-
-# PATCH_DIM = 64
-
-
-# def generate_full_resized_click_plot_data(
-#     img_data, click_coords, patch_size=(PATCH_DIM, PATCH_DIM)
-# ):
-#     img_shortlist = []
-#     img_titles = []
-#     img_coords = []
-#     for j in range(len(click_coords)):
-#         sliceno, y, x = click_coords[j]
-#         w, h = patch_size
-#         w += 10
-#         h += 10
-#         # print(x,y,w,h,sliceno)
-#         img_coords.append((sliceno, x, y, w, h))
-
-#         img = get_img_in_bbox(img_data, sliceno, int(np.ceil(x)), int(np.ceil(y)), w, h)
-#         img = resize(img, (28, 28))
-
-#         img_shortlist.append(img)
-#         # y_str = "{:.3f}".format(y)
-#         # x_str = "{:.3f}".format(x)
-#         img_titles.append(str(int(x)) + "_" + str(int(y)) + "_" + str(int(sliceno)))
-
-#     return img_shortlist, img_titles, img_coords
-
-
-# def get_single_class(
-#     class_string="DLP Full",
-#     window_sel=(0, 500),
-#     batch_size=32,
-#     offset=0,
-#     plot_all=False,
-#     plot_titles=True,
-#     patch_size=(PATCH_DIM, PATCH_DIM),
-# ):
-
-#     zoon_singleclass_df = zoon_anno_df.loc[
-#         zoon_anno_df["class_str"].isin([class_string])
-#     ]
-
-#     img_coords_df = zoon_singleclass_df[
-#         ["z", "x", "y", "class_code", "class_str", "xstart", "xend", "ystart", "yend"]
-#     ]
-#     img_coords_df["w"] = patch_size[0]
-#     img_coords_df["h"] = patch_size[1]
-
-#     img_coords_df = img_coords_df.iloc[window_sel[0] : window_sel[1]]
-#     # img_coords = img_coords.values
-
-#     singleclass_click_data = img_coords_df[["z", "x", "y"]]
-#     # singleclass_click_data=singleclass_click_data.iloc[window_sel[0]:window_sel[1]]
-#     print(singleclass_click_data.shape)
-
-#     p2_click_data_sel = np.array(singleclass_click_data.values).astype(np.float32)
-
-#     print(f"Selected single class data of shape {p2_click_data_sel.shape}")
-
-#     sel_clicks, x_coords_range, y_coords_range = generate_clicklist(
-#         p2_click_data_sel, crop_roi, 0, 165
-#     )
-
-#     print(f"Length of original click_data {sel_clicks.shape[0]})")
-
-#     vol_x_max = vol_shape_x - ((PATCH_DIM + 1) * 2)
-#     vol_x_min = (PATCH_DIM + 1) * 2
-
-#     vol_y_max = vol_shape_y - ((PATCH_DIM + 1) * 2)
-#     vol_y_min = (PATCH_DIM + 1) * 2
-
-#     vol_z_max = vol_shape_z - 20
-#     vol_z_min = 20
-
-#     print(vol_x_max, vol_x_min, vol_y_max, vol_y_min, vol_z_max, vol_z_min)
-
-#     out_of_bounds = np.hstack(
-#         (
-#             np.where(sel_clicks[:, 1] >= vol_x_max)[0],
-#             np.where(sel_clicks[:, 1] <= vol_x_min)[0],
-#             np.where(sel_clicks[:, 2] >= vol_y_max)[0],
-#             np.where(sel_clicks[:, 2] <= vol_y_min)[0],
-#             np.where(sel_clicks[:, 0] >= vol_z_max)[0],
-#             np.where(sel_clicks[:, 0] <= vol_z_min)[0],
-#         )
-#     )
-
-#     sel_clicks = np.delete(sel_clicks, out_of_bounds, axis=0)
-
-#     print("Length after deleting out of bounds clicks: {}".format(sel_clicks.shape[0]))
-
-#     win_start = window_sel[0]  # batch_size * offset
-#     win_end = window_sel[1]  # win_start + batch_size + 1
-
-#     prepared_images, img_titles, img_coords = generate_stacked_click_plot_data(
-#         (wf1, wf2), sel_clicks[win_start:win_end], patch_size=patch_size, resize=True
-#     )
-#     prepared_titles = [str(class_string) + "_" + img_title for img_title in img_titles]
-
-#     # img_class_str = np.array([class_string] * len(img_coords))
-#     # print(prepared_images[0])
-
-#     print("Length of image shortlist: {}".format(len(img_shortlist)))
-#     print(len(sel_clicks), x_coords_range, y_coords_range)
-#     print(vol_shape_x, vol_shape_y, vol_shape_z)
-
-#     grid_side_dim = np.int(math.sqrt(batch_size))
-#     print("Grid side dim: {}".format(grid_side_dim))
-
-#     if not plot_titles:
-#         img_titles = [0] * len(img_titles)
-
-#     # if plot_all:
-#     # plt.title(class_string)
-#     #   grid_of_images(selected_images[win_start:win_end], grid_side_dim,
-#     #                      grid_side_dim,  image_titles=img_titles)
-#     #
-#     return np.array(prepared_images), np.array(prepared_titles), np.array(img_coords_df)
-
-
-# def generate_click_plot_data1(img_data, click_coords):
-#     img_shortlist = []
-#     img_titles = []
-
-#     for j in range(len(click_coords)):
-
-#         if j % 5000 == 0:
-#             print("Generating click plot data: {}".format(j))
-
-#         sliceno, y, x = click_coords[j]
-#         w, h = (100, 100)
-#         print(x, y, w, h, sliceno)
-
-#         img = get_img_in_bbox(img_data, 75, int(np.ceil(x)), int(np.ceil(y)), w, h)
-#         img_shortlist.append(img)
-
-#         y_str = "{:10.4f}".format(y)
-#         x_str = "{:10.4f}".format(x)
-#         img_titles.append(x_str + " " + y_str + " " + "Slice no: " + str(sliceno))
-
-#     return img_shortlist, img_titles
