+++ survos2/frontend/plugins/analyzers/label_analysis.py
@@ -4,7 +4,7 @@ from matplotlib.backends.backend_agg import FigureCanvasAgg
 from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg
 from matplotlib.figure import Figure
 from napari.qt.progress import progress
-from survos2.entity.cluster.cluster_plotting import cluster_scatter, image_grid2, plot_clustered_img
+from survos2.entity.cluster.cluster_plotting import cluster_scatter, image_grid, plot_clustered_img
 
 from survos2.frontend.control import Launcher
 from survos2.frontend.components.base import (
